Fix issue #286
Flow Logs are not per-VPC resources from an API point of view, but rather tied to a region and the mapping between flow logs and VPCs/Subnets is indicated by the ResourceId attribute. This mapping is duplicated by the sort_vpc_flow_logs_callback, but the path to the flow_logs was wrong in the metadata config file. The logic for adding flow logs to VPCs/Subnets was buggy too.