Don't generate redshift findings for default parameter groups
AWS Redshift generates default parameter groups (e.g. "default.redshift-1.0")
that have certain parameters that can generate findings, such as
"enable_user_activity_logging" == "false", and "require_ssl" == "false".

The user is unable to modify these parameters, so in cases where a default
parameter group exists, findings will be generated which the user has no
ability to remediate.

The "redshift-parameter-group-logging-disabled" and
"redshift-parameter-group-ssl-not-required" rules are modified to only apply to
non-default parameter groups for which the user can be expected to remediate.