Adjust AWS "secrets" regular expressions
The existing regexes are prone to false positives because they match
strings which might be substrings of longer, innocuous strings.
"Secret Access Keys" in particular will be detected (in error) when
nearly any sufficiently long pathname is encountered.

This commit strengthens the check to ensure any potential match is
delimited (somehow), disallowing substring matches. False positives
still are possible, but the target string now must be EXACTLY the
right length, not at least the right length.

Signed-off-by: Scott Bailey <scott.bailey@godaddy.com>