Add unit tests for regex fix
At present, the new unit test fails. Bizarrely, the test works as
expected if it is performed manually instead of using pytest (i.e.
define SAMPLE_USER_DATA and then call ._identify_user_data_secrets()
with it. The regexes also work correctly when tested interactively or at
pythex.org.

It just doesn't work correctly for pytest (and by presumption, for
actual code -- I'm working to construct a test for that).