Disable and annotate new unit test
Good luck figuring this one out... I have verified actual behavior, and
interactive invocation (of both the individual commands, and the method)
work exactly as expected.

For some reason, using pytest to evaluate this configuration appears to
break proper parsing of the regular expressions (??!!?!) and the
assertions in the test fail because they return substring matches which
are explicitly prohibited by the final group of the regular expression.

I have annotated the test to indicate it is expected to fail, and added
documentation of actual "in the wild" behavior, which is as expected.