Disable KeyVault diagnostic settings checks
Disable checks related to KeyVault diagnostic settings since some of the functions involved do not work as expected. get_diagnostic_settings invokes the "list" operation from DiagnosticSettingsOperations for the MonitorManagementClient class, which expects a "resource_uri" parameter consisting on an Azure resource URI. This is not currently being passed onto that function correctly and generates errors.