Rule for ADA-CP 2.9.1 "Ensure IAM password policy prevents password reuse"
For this one, I modified the existing rule
"iam-password-policy-reuse-enabled".  Previously, that rule only checked
that there the account was configured to prevent reuse of some number of
previous passwords; I added a parameter to put a lower bound on the
number of remembered passwords.  This required me to modify the existing
rulesets; a parameter of 1 is equivalent to the old check.

I also modified the data model and rule display somewhat: ScoutSuite
previously synthesized a boolean to indicate whether password reuse
prevention was on in addition to storing the number of past passwords to
remember.  But AWS doesn't have that boolean; reuse prevention is
disabled if there is no number of past passwords to remember.  It only
allows 1-24 passwords to be remembered, so remembering 0 is equivalent
to having the feature disabled.  So I changed the data model to store 0
if it's disabled and check based on that.  This makes the rule logic and
display simpler.

While I was changing the data display, I found that output for the
settings controlling whether users are allowed to change their own
passwords was gated on password reuse prevention being enabled.  There
is no such dependency within AWS, so I removed that condition.