Rule for ADA-CP 2.10.1 "Ensure credentials unused for 45 days or greater are disabled"
This rule is closely related to "iam-unused-credentials-not-disabled".
However, that rule applies only to credentials that have been used; both
ADA-CP and CIS 2.0 also apply to credentials that have *not* been used.
I haven't checked earlier versions of CIS, so maybe they are more
limited in scope; alternately, "iam-unused-credentials-not-disabled" may
be just wrong and should be replaced with this new rule.